LayoutTests/imported/w3c/web-platform-tests/eventsource/format-field-parsing-expected.txt


FAIL EventSource: field parsing assert_equals: expected "\0\n 2\n1\n3\n\n4" but got "\0\n 2\n2\n1\n3\n\n4"
